[CH] getOldestNFiles now accepts longs instead of ints

+++ src/main/java/de/craftsblock/craftscore/utils/Utils.java
@@ -212,7 +212,7 @@ public final class Utils {
      * in ascending order by modification date.
      * @throws RuntimeException if an {@link IOException} occurs while accessing the folder.
      */
-    public static List<Path> getOldestNFiles(Path folder, int amount) {
+    public static List<Path> getOldestNFiles(Path folder, long amount) {
         try (Stream<Path> paths = Files.walk(folder)) {
             return paths.filter(Files::isRegularFile)
                     .sorted(Comparator.comparingLong(p -> p.toFile().lastModified()))
